+++ Material.Blazor/Components/Switch/MBSwitch.razor.cs
@@ -15,15 +15,7 @@ namespace Material.Blazor
         [Parameter] public string Label { get; set; } = "On/off";
 
 
-        /// <summary>
-        /// The switch's density.
-        /// </summary>
-        [Parameter] public MBDensity? Density { get; set; }
-
-
         private ElementReference ElementReference { get; set; }
-
-        private MBCascadingDefaults.DensityInfo DensityInfo => CascadingDefaults.GetDensityCssClass(CascadingDefaults.AppliedSwitchDensity(Density));
 
 
         // Would like to use <inheritdoc/> however DocFX cannot resolve to references outside Material.Blazor
@@ -32,12 +24,20 @@ namespace Material.Blazor
             await base.OnInitializedAsync();
 
             ConditionalCssClasses
-                .AddIf(DensityInfo.CssClassName, () => DensityInfo.ApplyCssClass)
-                .AddIf("mdc-switch--disabled", () => AppliedDisabled)
-                .AddIf("mdc-switch--checked", () => ComponentValue);
+                .AddIf("mdc-switch--unselected", () => !ComponentValue)
+                .AddIf("mdc-switch--selected", () => ComponentValue);
 
             SetComponentValue += OnValueSetCallback;
             OnDisabledSet += OnDisabledSetCallback;
+        }
+
+
+        /// <summary>
+        /// Toggles Value when the button is clicked.
+        /// </summary>
+        private void ToggleOnClick()
+        {
+            ComponentValue = !ComponentValue;
         }