+++ Material.Blazor/Model/MBEnumerations.cs
@@ -179,9 +179,19 @@ namespace Material.Blazor
         Mini,
 
         /// <summary>
-        /// FAB extended variant.
+        /// FAB extended variant without icon.
         /// </summary>
-        Extended
+        ExtendedNoIcon,
+
+        /// <summary>
+        /// FAB extended variant with leading icon.
+        /// </summary>
+        ExtendedLeadingIcon,
+
+        /// <summary>
+        /// FAB extended variant with trailing icon.
+        /// </summary>
+        ExtendedTrailingIcon,
     }
 
 
@@ -434,12 +444,12 @@ namespace Material.Blazor
         Regular,
 
         /// <summary>
-        /// Applies the mdc-list--dense CSS class.
+        /// Applies the mdc-deprecated-list--dense CSS class.
         /// </summary>
         Dense,
 
         /// <summary>
-        /// Applies the mdc-list--avatar CSS class.
+        /// Applies the mdc-deprecated-list--avatar CSS class.
         /// </summary>
         Avatar
     }